pyproject.toml
[tool.poetry]
name = "referal-system"
version = "0.1.0"
description = ""
authors = ["NASTY-SMIT <nasty.smit.2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.0"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
adrf = "^0.1.4"
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
